First commit
A unique repository for holding all the project.
In the AngularWebAppSPA there is the source code for the single web page application in Angular. In ExpressBackend there is the source code for the Node JS Express backend server. In PulumiOktaResourcesSetup there is the source code for the Okta resources creation with pulumi.